termux-app: Embed the proot loader
The proot loader is a statically linked executable and needs to be
embedded with the APK so that it can be executed.

It's small (around 5 kB) and haven't changed for 6 years, so this should
be fine. If it does change we can add a version detection in proot and
prompt users to update.

Corresponding termux-packages commit:
https://github.com/termux-play-store/termux-packages/commit/d296c6ac6c93bff0019abd3d4145f0615468773f