Fix TermuxFileReceiverActivity
Since we don't have the `SYSTEM_ALERT_WINDOW` granted on Google Play
(and it's good to manage without that if possible) we need to start the
Termux activity explicitly, instead of going through the Termux service.

Also do not do I/O on the UI thread when receiving files to avoid ANR
(Activity Not Responding) errors when receiving files and I/O is
slow (due to big files, slow content providers, or slow writes to
storage).